requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
stripe==11.4.1
python-docx==1.1.2
reportlab==4.2.5
boto3==1.35.92
botocore==1.35.92
